media/gpu/chromeos: Migrate GetOriginalFrame to tracking token
Today a GenericSharedMemoryID is used to identify frames in order to
retrieve the original frame from a frame pool or other frame source.
This ID either comes from the NativePixmapDmabuf ID, DMA buffer FD, or
GpuMemoryBuffer->GetID(). Use of GpuMemoryBuffer->GetId() is deprecated

This replaces the GenericSharedMemoryID with a base::UnguessableToken,
which comes from FrameResource::tracking_token().

This also fixes a build dependency issue in the media/mojo/clients for
the stable video decoder.

Bug: b:376526494
Change-Id: I36882c422af988c4a0b643153fc42c759bb41da6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6055306
Reviewed-by: vikas soni <vikassoni@chromium.org>
Commit-Queue: Nathan Hebert <nhebert@chromium.org>
Reviewed-by: Pilar Molina Lopez <pmolinalopez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406876}