[Media] Fix DataBufferTest.ReadingWriting
This patch fixes an MSan error on the ReadingWriting test
in the DataBufferTest suite.

This error occurred due to accessing unitialized memory. This patch
fixes the issue by only accessing the initialized portion of memory.

In order to avoid this in the future, the data() method now returns
a base::span that is truncated to only return the valid portion of the
buffer. The writable_data() method still returns the entire buffer.

Bug: 380774524
Change-Id: I282086a3b23bb750291bb69b03e67439afa9c96d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6176692
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406895}