[NTP][Enterprise] Set up nav throttle for catching auth redirect
Sets up a |NavigationThrottle| to catch redirects back from Microsoft's
login page from the login flow on the NTP. To ensure erroneous
navigations are not caught a few things are checked:
1. The user is using a 1P NTP.
2. The opener's last url is chrome-untrusted://ntp-microsoft-auth.
3. An expected Microsoft login url is in the redirect chain.
4. The URL we expect is being navigated to.

This redirects the URL instead to about:blank. A navigation from the web
to a chrome-untrusted:// page usually changes the |BrowsingInstance|
and |SiteInstance|, which stops the opener from detecting the
redirect. To work around this, the popup is redirected to about:blank
with source_site_instance and initiator_origin set to match the opener
iframe. This sets the popup to match the |BrowsingInstance| and
|SiteInstance| so that its url can be read by the opener again.

The URL fragment is copied over from the original redirect to
about:blank, since it is information required for authentication.

Bug: 386379980
Change-Id: I34facc29c8c4e00966e65bddd7d1618c3b8c8450
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6068495
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Paul Adedeji <pauladedeji@google.com>
Commit-Queue: Riley Tatum <rtatum@google.com>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406912}