[ProfilePicker][Glic] Implement the Learn More button
The Learn More text is currently part of a bigger string. Therefore the
click even is added dynamically when the page is loaded by connecting
the part of the text through the class tag to override the click event.

On that event, the handler is notified to load the last used profile,
open or use a browser with it and load the "Sign in to Chrome" page from
the Help Center.

The page cannot be loaded directly, since the Picker uses the System
Profile which cannot be used to open a Browser.

The picker is not expected to be closed after this action.

Bug: b:385726690
Change-Id: I79a311f0bb69f1db94f52368c68f066643d2ec2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6168854
Reviewed-by: Nicolas Dossou-Gbété <dgn@chromium.org>
Commit-Queue: Ryan Sultanem <rsult@google.com>
Cr-Commit-Position: refs/heads/main@{#1406918}