Prevent AppMenu from closing after bookmark DnDs
BookmarkMenuDelegate now supports updating menu views when there are
updates to the model, which means we don't have to close the menu on
each update anymore.

This CL prevents AppMenu from closing when bookmarks nodes are moved,
allowing the menu views to be updated by the delegate.

Change-Id: I7791b86ac33458e7fd96e93c20cc968450f0e0c2
Bug: 40868588
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6148304
Reviewed-by: Christopher Grant <cjgrant@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Commit-Queue: Kaan Alsan <alsan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406920}