[views-ax] Use cache & remove PrivacyScreenToastView::accessible_name_
This CL updates the accessible name pattern in
`ash::PrivacyScreenToastView`. Previously, this View was using a
member variable to store and update the accessible name for its
associated `TrayBubbleView`.

This CL removes that member variable and instead uses the new ViewsAX
cache to store the accessible name. The value of the name was
previously updated by `PrivacyScreenToastView::SetPrivacyScreenEnabled`
and now it will be updated by the caller of that function instead,
because the caller has access to the `TrayBubbleView` object.

This CL also adds a new unit test file in order to test the logic
added here.

This CL is part of the ViewsAX project:
https://docs.google.com/document/d/1Ku7HOyDsiZem1yaV6ccZ-tz3lO2XR2NEcm8HjR6d-VY/edit#heading=h.ke1u3utej413

Bug: 325137417
Change-Id: I0c99c950e8da0d3242543ace0719a452f392fe91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6168412
Commit-Queue: Evelynn Kaplan <evelynn.kaplan@microsoft.com>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406930}