Set SiteForCookies in ContextualSearchDelegateImpl::ResolveSearchTermFromContext
ResourceRequest.site_for_cookies defaults to an empty SiteForCookies.
When UCC launches, when users opt to restrict 3P cookies a request
with an empty SFC will have its cookies blocked by 3PC blocking.

This CL sets the SiteForCookies to the site of the request URL. This
will send cookies when 3PC blocking is enabled. It will also send
SameSite cookies.

If this request is meant to be cross-site or we do not want to send
cookies in this context, the change should be ignored. In that case,
we may want to change `credentials_mode` to `omit` to indicate that
the request explicitly does not intend to include cookies.

Bug: 324879760
Change-Id: Id894b4fa42ef080604ec2fdb1094299adfa670d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6164811
Commit-Queue: Dylan Cutler <dylancutler@google.com>
Reviewed-by: Gang Wu <gangwu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406937}