Use OnExtensionReady instead of OnExtensionInstalled.
OnExtensionInstalled is only called the first time the extension
is installed, while OnExtensionReady is called every time it's
re-loaded.

We need to know when it's ready on subsequent loads, so that
we can trigger installs of the natural voices automatically.

This doesn't fix the bug where preferences aren't saved, but
this is a necessary fix before we can preferences with newly
loaded natural voices.

Bug: 374169780
Change-Id: Iec60b0fca500a7955ed007145571226376da7d73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6171240
Reviewed-by: Kristi Saney <kristislee@google.com>
Commit-Queue: Lauren Winston <lwinston@google.com>
Cr-Commit-Position: refs/heads/main@{#1406965}