Set SiteForCookies in PPAPIDownloadRequest::SendRequest
ResourceRequest.site_for_cookies defaults to an empty SiteForCookies.
When UCC launches, when users opt to restrict 3P cookies a request
with an empty SFC will have its cookies blocked by 3PC blocking.

This CL sets the SiteForCookies to the site of the request URL. This
will send cookies when 3PC blocking is enabled. It will also send
SameSite cookies. This code uses the Safe Browsing cookie store, but
it can stil set SiteForCookies anyway.

If this request is meant to be cross-site or we do not want to send
cookies in this context, the change should be ignored. In that case,
we may want to change `credentials_mode` to `omit` to indicate that
the request explicitly does not intend to include cookies.

Bug: 324879760
Change-Id: Ie28aee7ea6f5b14964178cd1da7c34d6314d848d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6172363
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Commit-Queue: Dylan Cutler <dylancutler@google.com>
Cr-Commit-Position: refs/heads/main@{#1406967}