[E2E] Disable opt-out enforcement with edge to edge everywhere flag
Use EdgeToEdgeEverywhere to disable the opt-in in theme. Also adding a commandline switch flag, so we can observe the "broken" UI to show what
are the area that needs to be fixed when e2e everywhere is enabled.

This change affects all activities that are based on ChromeBaseAppCompatActivity, which is the main focus to address. Theme.BrowserUI.AlertDialog.NoActionBar will be address in a follow up.

Bug: 388555432
Change-Id: I13d5d90421914c174f80a7cfc7b9333bda3a4d24
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6168652
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Wenyu Fu <wenyufu@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1406995}