[freezing] Rename the origin trial opt-out to BackgroundPageFreezeOptOut.
To avoid reusing the name of an existing trial that was run in the past.

The new name correctly conveys that this opt-out applies to background
page freezing, but not to other types of freezing (e.g. for BFCache).

Bug: 325954772
Change-Id: I4fef2eace3f6f58533452a99f275e5dc22073ef0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6175636
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Kentaro Hara <haraken@chromium.org>
Auto-Submit: Francois Pierre Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1407006}