[pip] Notify java side `onActivityStopped` from native
When `onActivityStopped` is called in OverlayWindowAndroid(.cc),
previously it would call `Destroy()` without notifying the java
side.  Unfortunately, `Destroy()` explicitly assumes that the
java side has forgotten the native ptr, since it drops the java
ref.  The result is that the java side is never told that the
native side has been destroyed at some point later, and can
gleefully access the native side as a UAF if it gets `onDestroy`.

This CL removes the call to `Destroy()`, and instead chains the
`OnActivityStopped()` call to `Close()`.  This requests that the
java side clean up properly, including forgetting the native
pointer.  It also cleans up the compositor, so that it no longer
references the may-eventually-be-destroyed WindowAndroid.

This CL also renames `Destroy` so that it's clear that it should
be called from the java side only to notify the native side that
the native side should begin cleanup, in response to the java side
discovering that the pip window has closed / etc.

Bug: 389122107
Change-Id: I82ff401a1022f634cace42a63ddb0745131437af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6171221
Reviewed-by: Bo Liu <boliu@chromium.org>
Commit-Queue: Frank Liberato <liberato@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1407040}