[carousel] Select closest scroll-marker
The CSS overflow 5 spec[1] relaxed its scroll-marker selection
algorithm from being normative to only being a suggested example.
The algorithm in the spec's example, by selecting the "largest target
position where target position is equal to or before"[2] the scroller's
scroll position has the disadvantage, as pointed out in the GitHub
issue[3], that "even when the next section is 90%+ visible ... the
active marker does not update to this section."

This patch modifies Chromium's selection algorithm to instead consider
which scroll target's aligned position is closest to the scroller's
scroll position, which is similar to what is done for scroll snap
selection.

The WPT is added in wpt_internal instead of external/wpt because
the spec doesn't have a normative algorithm.

[1] https://drafts.csswg.org/css-overflow-5/#active-scroll-marker
[2] https://drafts.csswg.org/css-overflow-5/#:~:text=largest%20target%20position%20where%20target%20position%20is%20equal%20to%20or%20before
[3] https://github.com/w3c/csswg-drafts/issues/10917#:~:text=even%20when%20the%20next%20section%20is%2090,the%20active%20marker%20does%20not%20update%20to%20this%20section

Bug: 370092190,388256214
Change-Id: If117c5bcc7447487ea4f45d0ceb3d4bc9460a9ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6174056
Reviewed-by: Steve Kobes <skobes@chromium.org>
Commit-Queue: David Awogbemila <awogbemila@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1407087}