Fix cross-window focus
When a new window is created, the new GlicBorderViewManager will not
have the context sharing status state set because the WebApp won't call
SetContextAccessIndicator() for a focus change.

The solution is to let the border manager to actively query the latest
status at initialization. However GlicBorderViewManager is created
before the View tree is constructed [0] so even if GlicBorderViewManager
has the latest bool, it will need to wait for the BorderView to
initialize then update the border.

This CL consolidates the GlicBorderViewManager into BorderView to avoid
the wait, so that the border view gets updated when it's created in a
new Window.

Tested: browser tests; manually tested by adding
`SetContextAccessIndicator(true)` at the end of `LaunchUI()`.

[0]: https://crsrc.org/c/chrome/browser/ui/browser.cc;drc=c0883e36f0f65273f002c2ca8b7e9474256e00e4;l=660

Bug: 388901740
Change-Id: I283372c30f7a2e0515e4464dc51a6f40c41bcf80
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6175628
Reviewed-by: Ian Wells <iwells@chromium.org>
Commit-Queue: William Liu <liuwilliam@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1407119}