Fix the vectorization of encoded form and field feature tokens
The current implementation of padding in
FieldClassificationModelEncoder is leading to ASAN failures, because
despite the memory for the tokens vector is reserved, std::fill
still technically writes outside of vector bounds.

Fixing this by changing the way the padding tokens are inserted,
and changing the form encoding unittests to include the padding case.

Bug: 389873208
Change-Id: I441d60860d474cf4dd4ab09cd8e7e8c479130408
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6171722
Commit-Queue: Dominic Battré <battre@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Dominic Battré <battre@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1407153}