Implement No-Vary-Search in-memory store
Implement net::NoVarySearchCache, an in-memory mapping from URLs that
will be looked up in the cache to previously-fetched URLs that are
equivalent under the rules of No-Vary-Search (see
https://httpwg.org/http-extensions/draft-ietf-httpbis-no-vary-search.html).

This will be used to add No-Vary-Search support to
net::HttpCache::Transaction in a future CL.

This version does not yet support persistence of the cache to disk. That
will be implemented in a future version.

NoVarySearchCache will be owned by HttpCache, so there will be one
instance per URLRequestContext. The object will be created when a
Profile is created, and exist until it is destroyed (usually
browser shutdown).

See the design doc at
https://docs.google.com/document/d/1RS3q6qZ7-k9CvZsDYseGOXzcdQ9fGZ6YYnaW7fTPu7A/edit.

BUG=382394774
BYPASS_LARGE_CHANGE_WARNING=It is what it is
NO_IFTTT=I did change enums.xml but it complains anyway.

Change-Id: Ief53f1b4fd1f94366b3d719635d8b4fc553da5fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6088312
Commit-Queue: Adam Rice <ricea@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Auto-Submit: Adam Rice <ricea@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1407159}