NoVarySearchCache: Use same partitioning as HttpCache
Previously, NoVarySearchCache was partitioned by NetworkIsolationKey.
Unfortunately, the HttpCache has some additional values included in its
partition key in addition to the NetworkIsolationKey. Furthermore, these
values vary depending on the active experiment(s).

If the NoVarySearchCache is partitioned differently than the HttpCache
then there is a risk of leaking information across partitions, which
would defeat the security & privacy guarantees provided by cache
partitioning.

Use the same cache key generator for NoVarySearchCache as HttpCache to
ensure they are partitioned the same.

Unfortunately, this leads to a large increase in the number of string
allocations and copies needed to detect a cache miss. Follow-up work
will be needed to reduce this cost.

BUG=382394774

Change-Id: Ibd3c7b06868082e6e3a2e19d1d0e644e64ce4d7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6170844
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1407160}