[Uno-D] Fix "Signin to Chrome" button in settings.
This CL is similar to
[Uno-D] Fix "Continue as" button in settings | https://crrev.com/c/6157848

It fixes the same bug, but for another variant of the button.
Sync is now treated as a promo and can be declined.

Screencast: http://go/scrcast/NDc1NzAwMDEwODkwMDM1MnxmYTlhMzg4MC0zMw

Fixed: 387472002
Change-Id: I4df511f1da60364c19c08975c6ab5956f8d19138
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6176921
Reviewed-by: Liza Bipin <mlbipin@google.com>
Commit-Queue: David Roger <droger@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1407196}