Move the customizable select resources into the main stylesheets.
Now what we have @supports blink-feature(), there's no need to keep
these separate in code and add them after-the-fact lazily to the main
ruleset; it only adds complexity.

Ideally, we'd like to get rid of adding to RuleSet after-the-fact
entirely (it means we have to keep support for uncompacting them),
but we currently cannot do that due to printing. Also, we cannot
do this for PEPC since features in origin trials are not supported.

Bug: 41483940
Change-Id: I691290c5447f7f020d2045113edb0bc8f8157a31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6170442
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Auto-Submit: Steinar H Gunderson <sesse@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1407202}