Fix failing testCorrectSelectionMenuItemsAddedForPlainTextSelection test
This test was failing on Android Automotive OS and has been disabled.
Due to the limited apps on the OS, certain context menu actions
like share and websearch are not present on the menu which the test
was looking for. This CL adds a check to handle that.

Bug: 385155403
Change-Id: I1411af467ac852daa29a7f20e0378e429ebbc853
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6157849
Commit-Queue: Zakarie Yaris <yaris@google.com>
Reviewed-by: Alex Mitra <alexmitra@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1407227}