Replace raw_ptr/* to VideoCaptureController with scoped_refptr.
Raw pointers increase risk of UAF and begs lifetime questions. In order
to improve code health, we replace all raw pointers to the
VideoCaptureController with scoped_refptr.

This gives up some freedom as to when VideoCaptureController is
destroyed or on what thread this happened, since a scoped_refptr could
be bound on a task being posted. By adding some DCHECKs in its
destructor, we ensure this isn't a problem:
- We should already have "cleaned everything up" (e.g. removed clients
  and devices) by the time we release the last reference to the
  controller).

While this assumption appears to hold in practice (as per manually
running Chrome and opening and closing the camera), the unit tests were
not very good at cleaning things up so they are updated accordingly.

Bug: 385089402
Change-Id: Ib88a1b800220564258493866ce0f7d4ba5dbb960
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6110734
Reviewed-by: Ilya Nikolaevskiy <ilnik@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Commit-Queue: Henrik Boström <hbos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1407230}