WebView Prerender: Support errorCallback
This CL implements an initial support of errorCallback for prerendering.

In this initial version, the callback can be fired for any types of
prerender cancellation, ideally the callback should be called only when
prerendering is canceled for unexpected reasons though. The proper
support will be implemented in follow-up CLs.

Also, this CL passes a generic error message ("Prerendering fails")
instead of a specific error reason. This will be improved in the
follow-ups as well.

Bug: 41490450
Change-Id: Ie3ee3bcdc79ea0f6eb181419ec84e0eda829732c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6090374
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Reviewed-by: Lingqi Chi <lingqi@chromium.org>
Reviewed-by: Ziad Youssef <ziadyoussef@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1407241}