[ProfilePicker][Glic] Create a temporary testing switch flag
This code will be reverted when the Glic view will hook the call to the
Profile Picker Glic version as it won't be needed anymore. In the
meantime it allows, to easily test the Profile Picker Glic version.

Bug: b:390212241
Change-Id: Ifdc1834163e537855355f64f6b78a4679300065e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6171726
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Ryan Sultanem <rsult@google.com>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1407266}