Add HMAC secret generation
PRF support requires the use of an hmac secret, so this
CL adds the ability to generate the hmac secret to the
GeneratePasskeyAndEncryptSecrets function.

Also, passkey_model_utils_unittest.cc was not compiled
on iOS, so this CL adds it to the iOS build and tests.

Bug: 355045497
Change-Id: I31f1ae183c850eda48bac34434c4c56e0cede44f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6176488
Commit-Queue: Alexis Hétu <sugoi@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/main@{#1407285}