Create PageActionContainerView before the legacy PageActionIconContainer
This CL adds the new `PageActionContainerView` to the location bar
before `PageActionIconContainer`. This is desirable because the icons in
the former should appear to the _left_ of the icons in the latter. This
was affecting the order in which icons were focused; e.g., repeatedly
hitting tab while focused on the location bar would first focus the
legacy icons before the newer icons (jumping from right to left).

Bug: 376285664
Change-Id: Ia278a7ef89b72b26949bfb8aa137b51c3fd6f02b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6169661
Commit-Queue: Kaan Alsan <alsan@chromium.org>
Reviewed-by: Christopher Grant <cjgrant@chromium.org>
Reviewed-by: Caroline Rising <corising@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1407300}