[Android][3PPWM] Change to the flag protecting the cached 3P mode value
The content provider feature will be launched together with the virtual
view structure feature so setting the value that it provides is now
protected by both flags.

The content provider flag is on by default so it's easier to have the
feature tested using just the virtual view structure flag. The rationale
for default-enabling is that the content provider is externally accessible already due to the unguardable manifest entry.

The content provider entry in flag is not removed because we
still want to be able to roll back the content provider part feature.

Bug: 384485045
Change-Id: Ib11a2eb4e62548c27ee76b5f672ede349616fd46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6168865
Reviewed-by: Jan Keitel <jkeitel@google.com>
Commit-Queue: Ivana Žužić <izuzic@google.com>
Reviewed-by: Friedrich Hauser <friedrichh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1407325}