Allow tests to define static/constexpr GaiaId literals
Before this patch, there are about 500 casts in tests from a string
literal to GaiaId (and more to come on iOS, once implicit conversion
from std::string to GaiaId is removed).

This harms readability and also increases the risk for bugs in tests.

In this patch, a new test-only class is introduced that allows
defining constexpr/static GaiaId literals, in a way that it converts
implicitly to GaiaId. The idea that this implicit conversion is
allowed by the style guide has been discussed with cxx@chromium.org,
concluding that it is fine for this case.

The patch also adopts the class in a few tests but many other
occurrences are scoped out.

Change-Id: I6daa90e2f7d345ad054e07303e542c7672dde168
Bug: 380416867
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6165190
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1407377}